Dev to master (#31)
* implement transfer event catch

* implement getting balances endpoint using new table

* add new endpoints to fetch the balance

* up

* more

* up

* up

* up

* up

* up

* up

* up

* up

* up

* up

* 404

* up

* up

* up

* more leaderbork

* more

* typeahead

* up

* token owner search fix

* container

* up

* up

* go2sleep

* hack for pushing route on leaderbork

* more leaderbork

* up

* up

* up

* up

* up

* up

* up

* up

* up

* fix

* fix pixel preview offset

* fix

* fix

* correct endpoint

* check paths on building server

* dev deploy

* up

* up

* up

* up

* up

* upp

* up

* up

* fix

* fix2

* fix

* up

* fix

* up

* format

* get price

* drop scrolling header

* up

* up

* goodbye rinkeby you will be missed

* logs

* fix

* up

* fix

* fix

* up

* scale pixel preview on size change

* leaderbork mobile layout

* up

* up

* up

* up

* more

* up

* more

* up

* test burn and mint buttons

* up

* fix

* up

* more

* fixes

* up

* up

* mobile nav

* smooth mobile nav

* complete nav

* mobile nav

* fix search inputs

* test balance

* readme

* up

* fix

* up

* fix

* small fixes

* fix no UD

* uds in transfers

* up

+++ packages/server/dev-local.sh
@@ -7,15 +7,27 @@
 #   ./dev-local.sh up
 #   ./dev-local.sh down
 
+set -eu
+
+export SCRIPTPATH="$( cd "$(dirname "$0")" >/dev/null 2>&1 ; pwd -P )"
+
+
 spacedEcho() {
     echo ""
     echo "$1"
     echo ""
 }
 
-up() {
+removePreviousBuildMaybe() {
+  if [ -d "$SCRIPTPATH/dist" ]
+  then
     spacedEcho "removing old dist"
     rm -r dist/
+  fi
+}
+
+up() {
+    removePreviousBuildMaybe
 
     spacedEcho "spinning up db"
     docker-compose up -d db
@@ -29,7 +41,7 @@ up() {
     yarn compile_contracts
 
     spacedEcho "migrating local db"
-    yarn run prisma migrate dev --name init
+    yarn prisma:migratedev
 
     spacedEcho "spinning up api"
     if [[ $1 = "--build" ]]; then
@@ -62,11 +74,11 @@ HELP_USAGE
 if [[ $1 == "down" ]]; then
     down
 elif [[ $1 == "up" ]]; then
-    up $2
+    up ${2-""}
 else
     usage
 fi
 
-trap handler SIGINT
+trap handler SIGQUIT
 down
